Moved get_otu_accessions(), get_sequence_metadata() and get_otu_accessions_metadata() to .utils.reference
