+++ virtool_cli/utils/reference.py
@@ -1,6 +1,9 @@
+import json
 import re
 from pathlib import Path
 from typing import Optional, Tuple
+
+from virtool_cli.utils.storage import parse_sequence
 
 
 def is_v1(src_path: Path) -> bool:
@@ -119,3 +122,63 @@ async def get_unique_otu_ids(src_path: Path) -> list:
         if otu_path.is_dir()
     ]
     return unique_otus
+
+
+def get_otu_accessions(otu_path: Path) -> list:
+    """
+    Gets all accessions from an OTU directory and returns a list
+
+    :param otu_path: Path to an OTU directory under a reference directory
+    :return: A list of all accessions under an OTU
+    """
+    accessions = []
+
+    for isolate_path in get_isolate_paths(otu_path):
+        for sequence_path in get_sequence_paths(isolate_path):
+            sequence = json.loads(sequence_path.read_text())
+            accessions.append(sequence["accession"])
+
+    return accessions
+
+
+async def get_sequence_metadata(sequence_path: Path) -> dict:
+    """
+    Gets the accession length and segment name from a sequence file
+    and returns it in a dict
+
+    :param sequence_path: Path to a sequence file
+    :return: A dict containing the sequence accession, sequence length and segment name if present
+    """
+    sequence = await parse_sequence(sequence_path)
+
+    sequence_metadata = {
+        "accession": sequence["accession"],
+        "length": len(sequence["sequence"]),
+    }
+
+    segment = sequence.get("segment", None)
+    if segment is not None:
+        sequence_metadata["segment"] = segment
+
+    return sequence_metadata
+
+
+async def get_otu_accessions_metadata(otu_path: Path) -> dict:
+    """
+    Returns sequence metadata for all sequences present under an OTU
+
+    :param otu_path: Path to an OTU directory under a reference directory
+    :return: An accession-keyed dict containing all constituent sequence metadata
+    """
+    # get length and segment metadata from sequences
+    all_metadata = {}
+
+    for isolate_path in get_isolate_paths(otu_path):
+        for sequence_path in get_sequence_paths(isolate_path):
+            sequence_metadata = await get_sequence_metadata(sequence_path)
+
+            accession = sequence_metadata["accession"]
+
+            all_metadata[accession] = sequence_metadata
+
+    return all_metadata
